Add list of parameters for the acados matlab interface
The added table contains the list of parameters with the following columns: parameter name, description, accepted values and other info such as relevant references. It is not exhaustive yet.

The best workflow to generate the table that I found for now is the following:
1) edit the shared file: https://docs.google.com/document/d/14lLBsLHhSodpI9Lhq3YMoge5dXzUjT22pmzlMGeta50/edit?usp=sharing
2) download in .docx and convert to Github-flavored markdown with pandoc
3) copy-paste to "index.md"
Any different proposal is well accepted.