acados_template: add number of stages to be an argument to acados_create function
Since the model does not depend on the number of stages N, N is application dependent
and it should be possible to reuse the exported code in C with a different
number of stages N. This is made possible here.
As an example main.in.c and main_sim.in.c are modified accordingly.