fix: path constraints expression (h) is enforced at inital state
Bug description:
when using con_h_expr to set path constraints for an OCP, these constraints are enforced at nodes (0->N-1). However, typically, we also provide an initial state.
This results in having two constraints at the initial node. This could be a problem when both of them are active.

Fix:
- enforce path constraints at nodes (1->N-1)
- create a new variable (con_h_expr_0) to set a constraint expression at node 0 explicitly (in case it is desired)
- added a m inimal example to reproduce the issue and show usage of the new added variable (ocp_example_h_init_contraints.py)