Work on core for speedups (#888)
- **Cast highest level workspace (SQP/SQP-RTI) only in precompute for
more speed**:
This still allows to reuse workspace across modules. This can only break
if one tries to reuse the acados workspace memory from outside.
- **Add CasADi sparsity detection** that is called when creating an
external function.
The CasADi sparsity struct in the generated functions contains an `int`
which always say inputs are sparse, often this is not the case. Thus, we
check now for speed!
- Added `sim_in_assign_and_advance`, this can be called to avoid calling
`calculate_size` twice. Doing this everywhere is rather tedious and does
not save very much.
- **Breaking**: precompute is now required!

On a small benchmark of an RTI with ERK and the cartpole this gives me
>20% speedup, from 78μs to 62μs.