Funnel globalization (#1151)
This PR implements a new globalization method that is based on a funnel.
It can be used by setting the option `globalization` to
`FUNNEL_L1PEN_LINESEARCH` and is detailed below.
In addition, the PR contains the following general new options:
- `eval_residual_at_max_iter`: boolean that decides if the residuals
(and function and derivatives) will be evaluated in the last iteration.
The previous behavior of the SQP method is to _not_ evaluate again at
max iter. For the funnel methode, the default is true.
- `nlp_solver_tol_min_step_norm`: If this option is set to a value
larger than zero, the solver will terminate if a step size smaller than
the given value is encountered.

## Funnel method
A funnel method can be seen as a simplification of a filter method that
can achieve similar performance in terms of robustness. The implemented
algorithm uses a line search for step size determination. Additionally,
a penalty phase is implemented that should handle infeasible subproblems
similiarly to a feasibility restoration as, e.g., implemented in IPOPT.
The work on the funnel is subject of current research. Therefore, we are
currently working on an article that describes the algorithm in detail.

This PR includes the following features:
- new globalization strategy based on a funnel: The funnel is an upper
bound on constraint violation, that drives globalization. In case no
progress can be made, a penalty phase is incorporated that avoids a
feasibility restoration phase.
- an improvement of the termination criterion. Unbounded problems can be
detected, as well as when the algorithm converges to suboptimal point
and the step size goes below a given threshold.
- The iteration output for the `FUNNEL_L1PEN_LINESEARCH` includes more
data about the globalization parameters.
- The current `FUNNEL_L1PEN_LINESEARCH` globalization is tested on
feasibility problems and on some simple OCPs.

The following new funnel related options are added:
- `funnel_initialization_increase_factor`: initialization of funnel
width: `max(funnel_initialization_upper_bound,
funnel_initialization_increase_factor * initial_infeasibility)`
- `funnel_initialization_upper_bound`: see line above
- `funnel_sufficient_decrease_factor`: h-type steps need to declare
sufficient decrease in constraint violation wrt funnel
- `funnel_kappa`: convex combination factor for funnel decrease function
- `funnel_fraction_switching_condition`: fraction in switching condition
used for distinguishing between f and h steps
- `funnel_initial_penalty_parameter`: initialization of penalty
parameter, used in penalty phase that replaces a feasibility
restoration. This penalty parameter will later be used in l1 relaxation.

Disclaimer: The current funnel implementation is still preliminary. For
example, infeasible QPs cannot be handled yet which can mitigate the
convergence behavior of certain problems.

---------

Co-authored-by: Jonathan Frey <jonathanpaulfrey@gmail.com>
Co-authored-by: sandmaennchen <katrin.baumgaertner@posteo.de>