Migrate sensitivities of solution and value function of parameters to `p_global` (#1316)
Solution sensitivities with respect to parameters and derivatives of the
value function with respect to parameters are now computed with respect
to `p_global` instead of `p`. This is anyway what was implemented
conceptually. However, if different parameter values for stages were
used, these sensitivities were not clearly defined.
The precomputations and the fact that the parameter vector is only
defined once makes this implementation more efficient.

- added support for `p_global` in Cython

Breaking:
- concerns options `with_solution_sens_wrt_params`,
`with_solution_sens_wrt_params`
- `eval_solution_sensitivity` with argument `with_respect_to =
"params_global"`
- `eval_and_get_optimal_value_gradient` with argument `with_respect_to =
"params_global"`
- rename option `with_respect_to='params_global'` to `'p_global'` with
deprecation warning

---------
    
Co-authored-by: sandmaennchen <katrin.baumgaertner@posteo.de>