New QP diagnostics function in `acados_ocp_solver.py` and `adaptive_levenberg_marquardt` in multi phase interface (#1338)
First New added feature:
- `qp_diagnostics` function in `acados_ocp_solver.py`: this function
returns a dictionary with the min/max eigenvalues and the condition
number for the whole Hessian matrix as well as for every stagewise
Hessian. You can also add a flag `hessian_type` for either the
`FULL_HESSIAN` or `PROJECTED_HESSIAN`
- removed `hessian_eigenvalues` function in `sensitivity_utils.py`
- adjusted `qp_diagnostics` within Matlab `AcadosOcpSolver.m` so that it
behaves similarly to the python version. The projected Hessian feature
is still missing in the Matlab version

Second New added feature:
- `adaptive_levenberg_marquardt` options were added for multi phase as
well