Evaluator (#1402)
Adding a basic evaluator that can be used to (I) evaluate closed-loop
cost aligned with the acados definition and (II) to evaluate the full
trajectory obtained from acados externally. Point two is mainly used for
testing the one-step evaluation of (I).

Updating parameters in the solver after instantiation requires to call
the "update_all" function of the evaluator to align the evaluator and
the solver. Attention: Currently only parameters "p" and "p_global" are
updated. Other (bounds, weights, and references) will follow.

Example minimal_example_evaluation.py tests the function on a pendulum.

The evaluator is in a preliminary phase and requires updates to be fully
aligned with acados. Next steps:
- Implement the possibility of updating the reference
- Test multiphase formulations
- Implement scaling
- Implement convex over nonlinear costs

---------

Co-authored-by: Jonathan Frey <jonathanpaulfrey@gmail.com>
Co-authored-by: sandmaennchen <katrin.baumgaertner@posteo.de>