Work on sensitivity solver: add `setup_qp_matrices_and_factorize` and more (#1432)
### Added `setup_qp_matrices_and_factorize()`:
```
This function sets up the QP and factorizes Hessian matrix.

The current NLP iterate is used to setup the QP iterate, which includes t-slacks.
The t and lambda values are clipped according to `solution_sens_qp_t_lam_min` to avoid ill-conditioning.
Then the Hessian matrix is factorized.

If a two-solver approach is used to obtain solution sensitivities from acados, this function should be called before calling `eval_solution_sensitivity`, or `eval_adjoint_solution_sensitivity()`.

This is only implemented for HPIPM QP solver without condensing.
```

### Key differences of using `setup_qp_matrices_and_factorize()` vs.
`solve()` in the two-solver approach:
- The QP solver does not need to perform multiple IP iterations to
converge and only a single Riccati factorization is performed at the
current iterate.
- In the chain example this reduces the "Median time to factorize exact
Hessian QP:" 450.000 ms to 117.000 ms -> roughly 4x faster :rocket:
- In the `solve()`-approach, the Riccati factorization at the second
last QP iterate is used, as the system is not factorized at the final QP
iterate. However, for the final iterate the system of equations becomes
typically ill-conditioned, which is why the final iterate is modified,
by clipping the inequality multipliers and t slacks, see option
`solution_sens_qp_t_lam_min`.
- Less options that have to be set. The options that were needed to use
the `solve` approach are now commented:
```
        # Old settings needed, when calling solve() instead of setup_qp_matrices_and_factorize()
        # ocp.solver_options.globalization_fixed_step_length = 0.0 # to not perfrom an SQP step update
        # ocp.solver_options.nlp_solver_max_iter = 1
        # ocp.solver_options.qp_solver_iter_max = 200
        # ocp.solver_options.tol = 1e-10  # to "force" a QP solve
```

### How it works
The option `update_fact_exit` was added in HPIPM. This option being
active, plus the warm start, plus limiting the QP solver iterations to
0, results in the correct behavior.

### Other things that changed in this PR
- fixed warm-starting QP from NLP iterate, was not correctly implemented
in #1428, as the QP iterate needs to be condensed and eliminated degrees
of freedom need to be taken into account. Only for N==N2 condensing
right now, to be extended https://github.com/acados/acados/issues/1433