Dynamic num threads in batch solvers (#1483)
Previously, the number of threads to be used in the
`AcadosOcpBatchSolver` or `AcadosSimBatchSolver` batched methods had to
be set in the respective `AcadosOcpOptions` or `AcadosSimOptions` and
could not be changed without rebuilding the solver.
Now, it is a property that can be set dynamically, even after building
the solver.

The property `num_threads_in_batch_solve` in both `AcadosOcpOptions` and
`AcadosSimOptions` is now deprecated,
and warnings will be printed if it is used.
Instead, the `AcadosOcpBatchSolver` and `AcadosSimBatchSolver` have a
new (settable) property `num_threads_in_batch_solve` and take the
initial value in their constructor via a new
`num_threads_in_batch_solve` argument.
The value of the new attribute is now used to determine how many threads
the batched methods should use.

The property `num_threads_in_batch_solve` in `AcadosOcpOptions` or
`AcadosSimOptions` was previously used for determining whether the
solvers should be build with openmp and how the C code of the batched
methods should be templated (if `num_threads_in_batch_solve > 1`).
Instead a new property `with_batch_functionality` is used as a flag to
determine whether the solvers should be build with openmp and whether
the C code of the batched methods should be created at all.
This new property is not required to be set by the user himself, but is
being set by the BatchSolvers themselves by default (a warning is
printed though, to notify the user that the code is still build with the
openmp flag).

The matlab templates and the batch solver examples have been adjusted
accordingly. Further, the minimal batch solver examples include new
simple tests for the new `num_threads_in_batch_solve` property.

---------

Co-authored-by: Jonathan Frey <jonathanpaulfrey@gmail.com>