Python: Add `AcadosCasadiOcpSolver` & minor fixes (#1523)
The class `AcadosCasadiOcpSolver` was added, for now is not full
featured, e.g. slacks are missing.
The class aims to provide a very similar API as `AcadosOcpSolver` and
should allow for sanity checks, debugging, benchmarking and more.

Additional changes:
- added `create_default_initial_iterate` to create an `AcadosOcpIterate`
without creating a solver
- fixed some issues in the evaluator class
- fixed issue in C interface which made the algorithm fail when loading
an iterate with empty z and ERK

Follow-up tasks tracked in https://github.com/acados/acados/issues/1525