Implement Anderson acceleration for fixed step SQP-type algorithms (#1521)
Only take 1 pervious step into account, as in this case there is an
explicit formula to compute the step.
This formula is e.g. given in eq. (4.4) in [Pollock2021 -- Anderson
acceleration for contractive and noncontractive
operators](https://academic.oup.com/imajna/article/41/4/2841/6065020)
This feature can be used by setting `with_anderson_acceleration = True`.

A plot of the SCQP benchmark example is below:

![anderson_scqp_convergence](https://github.com/user-attachments/assets/68e3d04b-dc5b-4912-9b20-ee1b234855e5)

Plots corresponding to the convergence experiment with the Furuta
pendulum are below:

![contraction_rates_furuta_pendulum](https://github.com/user-attachments/assets/67959d74-16ba-450f-919e-aeb9a60456e7)

![convergence_furuta_pendulum](https://github.com/user-attachments/assets/82a8924f-455c-4737-928b-a8137883f852)