Ros2 Node Generation (#1622)
A generic Ros2 Humble Node generation for a simple MPC including generic
messages.
Create your `AcadosOcp` and add the `AcadosOcpRosOptions` class to the
ocp variable `ros_opts` as below:
```python
from acados_template import AcadosOcpRosOptions, AcadosOcp

ocp = AcadosOcp()
...

ocp.ros_opts = AcadosOcpRosOptions()
ocp.ros_opts.package_name = "ocp_pkg" 
```

That's it. Two packages are created next to the `c_generated_code`
folder. One which includes the interface for the messages (here
`ocp_pkg_interface`) to communicate with the main package, and the main
package with MPC functionality (here `ocp_pkg`).
All weights and constraints are implemented as node parameters, which
can be changed during runtime. The change of `ts` only changes the
control loop timer and not yet the `ts` of the OCP solver. The
references, OCP parameters as well as state can be updated via the given
message which are included in the interface package.

Look at the generated `Redme.md` in the main folder to see how to build,
start and test the generated packages.

This MPC is triggered by a timer, other versions like topic trigger are
not implemented yet.
Other architecture types like lifecycle nodes are also not implemented
yet as well as services or actions (other form of messages).

Similarly, a ROS2 node for the `AcadosSimSolver` can be generated.
Additionally, some improvements to the CI workflow are in this PR.

---------

Co-authored-by: Jonathan Frey <jonathanpaulfrey@gmail.com>