Backward compatibility of tera renderer (#1635)
Two issues:
1) older Linux versions with default tera binaries have issues where
libg6 is not found. To fix this. Run the following in a Python file:
```
acados_template import get_tera
tera_version = '0.0.34', force_download=True)
```
2) ROS templates are not compatibile with old tera version. Only
relevant if you try generating a ROS node.

This PR
- moved ROS templates to separate folder, to avoid loading them when not
needed. There is some logic in there that is not compatible with the old
tera renderer version (0.0.34).
- adapt `get_tera` to be able to download older version.
- print those known issues in case rendering fails.