Update detect_constraint_structure.m
Removed a redundant and incorrect assignment of g section constraints in the path constraints assignments.  The C, D, lg, and, ug are already assigned in an identical if() case some some lines previous (lines 215-221).  The removed lines performed these same assignments but C and D were assigned to invalid locations.  Based on the way the code is written, I assume this is left over from before version acados v0.4.0.